ledger-on-sql: Move database transaction methods to `Database`.
