+++ ledger/ledger-on-sql/src/main/scala/com/daml/ledger/on/sql/SqlLedgerReaderWriter.scala
@@ -15,7 +15,7 @@ import com.daml.ledger.on.sql.queries.Queries.Index
 import com.daml.ledger.participant.state.kvutils.DamlKvutils.{
   DamlLogEntryId,
   DamlStateKey,
-  DamlStateValue,
+  DamlStateValue
 }
 import com.daml.ledger.participant.state.kvutils.api.{LedgerReader, LedgerRecord, LedgerWriter}
 import com.daml.ledger.participant.state.kvutils.{Envelope, KeyValueCommitting}
@@ -24,19 +24,16 @@ import com.digitalasset.daml.lf.data.Ref
 import com.digitalasset.daml.lf.data.Time.Timestamp
 import com.digitalasset.daml.lf.engine.Engine
 import com.digitalasset.ledger.api.health.{HealthStatus, Healthy}
+import com.digitalasset.logging.LoggingContext
 import com.digitalasset.logging.LoggingContext.withEnrichedLoggingContext
-import com.digitalasset.logging.{ContextualizedLogger, LoggingContext}
 import com.digitalasset.platform.akkastreams.dispatcher.Dispatcher
 import com.digitalasset.platform.akkastreams.dispatcher.SubSource.RangeSource
 import com.digitalasset.resources.ResourceOwner
 import com.google.protobuf.ByteString
-import javax.sql.DataSource
 
 import scala.collection.JavaConverters._
 import scala.collection.immutable.TreeSet
-import scala.concurrent.duration.Duration
 import scala.concurrent.{ExecutionContext, Future}
-import scala.util.control.NonFatal
 
 class SqlLedgerReaderWriter(
     ledgerId: LedgerId = Ref.LedgerString.assertFromString(UUID.randomUUID.toString),
@@ -49,8 +46,6 @@ class SqlLedgerReaderWriter(
     logCtx: LoggingContext,
 ) extends LedgerWriter
     with LedgerReader {
-
-  private val logger = ContextualizedLogger.get(this.getClass)
 
   private val engine = Engine()
 
@@ -66,7 +61,7 @@ class SqlLedgerReaderWriter(
       .startingAt(
         offset.getOrElse(StartOffset).components.head,
         RangeSource((start, end) => {
-          val result = inDatabaseReadTransaction(s"Querying events [$start, $end[ from log") {
+          val result = database.inReadTransaction(s"Querying events [$start, $end[ from log") {
             implicit connection =>
               queries.selectFromLog(start, end)
           }
@@ -88,22 +83,23 @@ class SqlLedgerReaderWriter(
           .getOrElse(throw new IllegalArgumentException("Not a valid submission in envelope"))
         val stateInputKeys = submission.getInputDamlStateList.asScala.toSet
         val entryId = allocateEntryId()
-        val newHead = inDatabaseWriteTransaction("Committing a submission") { implicit connection =>
-          val stateInputs = readState(stateInputKeys)
-          val (logEntry, stateUpdates) =
-            KeyValueCommitting.processSubmission(
-              engine,
-              entryId,
-              currentRecordTime(),
-              LedgerReader.DefaultConfiguration,
-              submission,
-              participantId,
-              stateInputs,
-            )
-          queries.updateState(stateUpdates)
-          val latestSequenceNo =
-            queries.insertIntoLog(entryId, Envelope.enclose(logEntry))
-          latestSequenceNo + 1
+        val newHead = database.inWriteTransaction("Committing a submission") {
+          implicit connection =>
+            val stateInputs = readState(stateInputKeys)
+            val (logEntry, stateUpdates) =
+              KeyValueCommitting.processSubmission(
+                engine,
+                entryId,
+                currentRecordTime(),
+                LedgerReader.DefaultConfiguration,
+                submission,
+                participantId,
+                stateInputs,
+              )
+            queries.updateState(stateUpdates)
+            val latestSequenceNo =
+              queries.insertIntoLog(entryId, Envelope.enclose(logEntry))
+            latestSequenceNo + 1
         }
         dispatcher.signalNewHead(newHead)
         SubmissionResult.Acknowledged
@@ -127,50 +123,6 @@ class SqlLedgerReaderWriter(
       .selectStateByKeys(stateInputKeys)
       .foldLeft(builder)(_ += _)
       .result()
-  }
-
-  private def inDatabaseReadTransaction[T](message: String)(
-      body: Connection => T,
-  )(implicit logCtx: LoggingContext): T = {
-    inDatabaseTransaction(message, database.readerConnectionPool)(body)
-  }
-
-  private def inDatabaseWriteTransaction[T](message: String)(
-      body: Connection => T,
-  )(implicit logCtx: LoggingContext): T = {
-    inDatabaseTransaction(message, database.writerConnectionPool)(body)
-  }
-
-  private def inDatabaseTransaction[T](
-      message: String,
-      connectionPool: DataSource,
-  )(
-      body: Connection => T,
-  )(implicit logCtx: LoggingContext): T = {
-    val connection =
-      time(s"$message: acquiring connection")(connectionPool.getConnection())
-    time(message) {
-      try {
-        val result = body(connection)
-        connection.commit()
-        result
-      } catch {
-        case NonFatal(exception) =>
-          connection.rollback()
-          throw exception
-      } finally {
-        connection.close()
-      }
-    }
-  }
-
-  private def time[T](message: String)(body: => T)(implicit logCtx: LoggingContext): T = {
-    val startTime = System.nanoTime()
-    logger.trace(s"$message: starting")
-    val result = body
-    val endTime = System.nanoTime()
-    logger.trace(s"$message: finished in ${Duration.fromNanos(endTime - startTime).toMillis}ms")
-    result
   }
 }
